requirements.txt
altgraph==0.17.3
arrow==1.2.3
asgiref==3.5.2
auto-py-to-exe==2.26.1
binaryornot==0.4.4
bottle==0.12.23
bottle-websocket==0.2.9
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cookiecutter==2.1.1
cryptography==39.0.1
cycler==0.11.0
defusedxml==0.7.1
Django==4.0.9
django-browser-reload==1.6.0
django-cors-headers==3.13.0
django-tailwind==3.4.0
djangorestframework==3.12.4
Eel==0.14.0
fonttools==4.38.0
future==0.18.2
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.1
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
packaging==23.0
pefile==2022.5.30
Pillow==9.4.0
pycparser==2.21
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
PyJWT==2.6.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.7.1
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
text-unidecode==1.3
tzdata==2022.6
urllib3==1.26.12
whichcraft==0.6.1
zope.event==4.6
zope.interface==5.5.2
